remove add experience dialog, consolidate add experience string
