+++ apps/web/src/pages/Profile/CareerTimelineAndRecruitmentPage/CareerTimelineAndRecruitmentPage.stories.tsx
@@ -1,12 +1,10 @@
 import React from "react";
 import { ComponentStory, ComponentMeta } from "@storybook/react";
 
-import { OverlayOrDialogDecorator } from "@gc-digital-talent/storybook-helpers";
 import { fakeSkills, fakeExperiences } from "@gc-digital-talent/fake-data";
 import { notEmpty } from "@gc-digital-talent/helpers";
 
 import CareerTimelineAndRecruitment from "./components/CareerTimelineAndRecruitment";
-import AddExperienceDialog from "./components/AddExperienceDialog";
 
 export default {
   title: "Pages/Career timeline and recruitment",
@@ -20,11 +18,6 @@ const CareerTimelineAndRecruitmentTemplate: ComponentStory<
 > = (args) => {
   return <CareerTimelineAndRecruitment {...args} />;
 };
-const AddExperienceDialogTemplate: ComponentStory<
-  typeof AddExperienceDialog
-> = (args) => {
-  return <AddExperienceDialog {...args} />;
-};
 
 export const NoExperiences = CareerTimelineAndRecruitmentTemplate.bind({});
 export const WithExperiences = CareerTimelineAndRecruitmentTemplate.bind({});
@@ -32,7 +25,6 @@ export const NoExperiencesMissingSkills =
   CareerTimelineAndRecruitmentTemplate.bind({});
 export const WithExperiencesMissingSkills =
   CareerTimelineAndRecruitmentTemplate.bind({});
-export const AddExperienceDialogOpen = AddExperienceDialogTemplate.bind({});
 
 const mockExperiences = fakeExperiences(10);
 const mockExperienceSkills = mockExperiences
@@ -69,9 +61,4 @@ WithExperiencesMissingSkills.args = {
     requiredSkills: mockRequiredSkills,
     optionalSkills: mockOptionalSkills,
   },
-};
-
-AddExperienceDialogOpen.decorators = [OverlayOrDialogDecorator];
-AddExperienceDialogOpen.args = {
-  defaultOpen: true,
 };
